Refactoring upper/lower suffixes
